+++ src/usdb_syncer/db/sql/upsert_resource.sql
@@ -20,4 +20,5 @@ UPDATE
 SET
     fname = :fname,
     mtime = :mtime,
+    status = :status,
     resource = :resource